🔥fix: NutrientInfo 속성 type 수정

+++ android-app/app/src/main/java/com/swallaby/foodon/domain/food/model/MealInfo.kt
@@ -8,10 +8,10 @@ data class MealInfo(
     val imageFileName: String = "",
     val mealTime: String = "",
     val mealTimeType: String = "",
-    val totalCarbs: Int = 0,
-    val totalFat: Int = 0,
+    val totalCarbs: Double = 0.0,
+    val totalFat: Double = 0.0,
     val totalKcal: Int = 0,
-    val totalProtein: Int = 0,
+    val totalProtein: Double = 0.0,
     val mealItems: List<MealItem> = emptyList(),
 )
